AMP-3457
git-svn-id: file:///svn/trunk/amp@5493 9e8d75aa-d0cb-44a9-8341-a4b8ba1a38b4

+++ WEB-INF/src/org/digijava/module/aim/action/AddNewTheme.java
@@ -37,9 +37,9 @@ public class AddNewTheme extends Action
 		HttpSession session = request.getSession();
 		themes = ProgramUtil.getParentThemes();
 		themeForm.setThemes(themes);
-		themeForm.setProgramExternalFinancing((float)0.0);
-		themeForm.setProgramInernalFinancing((float)0.0);
-		themeForm.setProgramTotalFinancing((float)0.0);
+		themeForm.setProgramExternalFinancing(Double.valueOf(0));
+		themeForm.setProgramInernalFinancing(Double.valueOf(0));
+		themeForm.setProgramTotalFinancing(Double.valueOf(0));
 		//themeForm.setProgramTypeNames(ProgramUtil.getProgramTypes());
 		return mapping.findForward("forward");
 	}
